+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.4.0
 
 * for elasticsearch 7 and 8, query both `_template` and `_index_template` endpoints, fixes [#281](https://github.com/cars10/elasticvue/issues/281)
+* improve shards view
+* fix issue when multiple indices are selected and one is manually deleted
 
 ## 1.3.1